feat(#S11P21D109-284): 거래내역 구현

+++ backend/spring-server/src/main/java/com/semonemo/spring_server/domain/coin/dto/response/CoinResponseDto.java
@@ -5,7 +5,7 @@ import java.util.List;
 
 public record CoinResponseDto (
     Long userId,
-    BigInteger coinBalance,
-    BigInteger payableBalance
+    Long coinBalance,
+    Long payableBalance
 ) {
 }
